api/common/security.py
from passlib.context import CryptContext
from jose import jwt, JWTError
from datetime import datetime, timedelta
from api.common.config import settings

pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")
ALGORITHM = "HS256"

# Password helpers


def hash_password(password: str) -> str:
    return pwd_context.hash(password)


def verify_password(password: str, hash: str) -> bool:
    return pwd_context.verify(password, hash)

# JWT helpers


def create_jwt_token(user_id: str, expires_delta: int = None) -> str:
    user_id_str = str(user_id)
    expire = datetime.utcnow() + timedelta(seconds=expires_delta or 900)
    to_encode = {"sub": user_id_str, "exp": expire}
    return jwt.encode(to_encode, settings.JWT_SECRET, algorithm=ALGORITHM)


def create_refresh_token(user_id: str) -> str:
    user_id_str = str(user_id)
    expire = datetime.utcnow() + timedelta(days=7)
    to_encode = {"sub": user_id_str, "exp": expire, "type": "refresh"}
    return jwt.encode(to_encode, settings.JWT_SECRET, algorithm=ALGORITHM)


def decode_jwt_token(token: str) -> dict:
    return jwt.decode(token, settings.JWT_SECRET, algorithms=[ALGORITHM])
